benchmark/exercises/mountain_jump_tracker/examples/mountain_range.py
class MountainRange:
    def __init__(self, heights, max_jump):
        self.heights = heights
        self.max_jump = max_jump

    def can_complete_mission(self):
        jumps_used = 0
        for i in range(1, len(self.heights)):
            jumps_used += abs(self.heights[i] - self.heights[i-1])
            if jumps_used > self.jump_capacity:
                return False
        return True

    def validate(self):
        if not all(isinstance(height, int) for height in self.heights):
            return False, "All heights must be integers."

        if not 0 < self.max_jump <= 100:
            return False, "Max jump must be between 1 and 100."

        return True, ""

    def jumps_required(self):
        jumps = 0
        i = 0
        n = len(self.heights)

        while i < n - 1:
            for j in range(i + 1, i + self.max_jump + 1):
                if j >= n:
                    break
                if self.heights[j] >= self.heights[i]:
                    i = j
                    jumps += 1
                    break
            else:
                return "Not possible to cross the range"

        return jumps

    def most_difficult_jump(self):
        max_diff = 0
        for i in range(len(self.heights) - 1):
            diff = abs(self.heights[i + 1] - self.heights[i])
            if diff > max_diff:
                max_diff = diff
        return max_diff